+++ packages/inline-choice/src/index.js
@@ -1,4 +1,4 @@
-import Main from './main.jsx';
+import InlineChoice from './inline-choice';
 import React from 'react';
 import ReactDOM from 'react-dom';
 import {
@@ -13,7 +13,7 @@ export default class RootInlineChoice extends HTMLElement {
     this._session = null;
     this._rerender = () => {
       if (this._model && this._session) {
-        let elem = React.createElement(Main, {
+        let elem = React.createElement(InlineChoice, {
           model: this._model,
           session: this._session,
           onChoiceChanged: this._handleChoiceChange.bind(this)
@@ -28,7 +28,7 @@ export default class RootInlineChoice extends HTMLElement {
     this.dispatchEvent(
       new ModelSetEvent(
         this.tagName.toLowerCase(),
-        this.session && !!this.session.selectedChoice,
+        this.session && !!this.session.value,
         !!this._model
       )
     );
@@ -38,12 +38,6 @@ export default class RootInlineChoice extends HTMLElement {
 
   set session(s) {
     this._session = s;
-    this.dispatchEvent(
-      new SessionChangedEvent(
-        this.tagName.toLowerCase(),
-        this.session && !!this.session.selectedChoice
-      )
-    );
     this._rerender();
   }
 
@@ -52,11 +46,11 @@ export default class RootInlineChoice extends HTMLElement {
   }
 
   _handleChoiceChange(selectedChoice) {
-    this.session.selectedChoice = selectedChoice;
+    this.session.value = selectedChoice;
     this.dispatchEvent(
       new SessionChangedEvent(
         this.tagName.toLowerCase(),
-        this.session && !!this.session.selectedChoice
+        this.session && !!this.session.value
       )
     );
     this._rerender();