+++ packages/ruler/demo/config.js
@@ -1,0 +1,15 @@
+module.exports = {
+  elements: {
+    'pie-ruler': '..'
+  },
+  models: [
+    {
+      element: 'pie-ruler',
+      id: '1',
+      width: 400,
+      height: 100,
+      units: 10,
+      measure: 'imperial'
+    }
+  ]
+};