+++ packages/inline-choice/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="4.0.1"></a>
+## [4.0.1](https://github.com/pie-framework/pie-ui/compare/@pie-ui/inline-choice@4.0.0...@pie-ui/inline-choice@4.0.1) (2018-04-25)
+
+
+### Bug Fixes
+
+* **session:** rename var in root from 'selectedChoice' to 'value' ([9abb90d](https://github.com/pie-framework/pie-ui/commit/9abb90d))
+
+
+
+
+<a name="4.0.0"></a>
+# [4.0.0](https://github.com/pie-framework/pie-ui/compare/@pie-ui/inline-choice@3.0.1...@pie-ui/inline-choice@4.0.0) (2018-04-25)
+
+
+### Features
+
+* **demo:** add demo ([043495c](https://github.com/pie-framework/pie-ui/commit/043495c))
+* **layout:** inline-choice now uses an inline layout. ([7df1d16](https://github.com/pie-framework/pie-ui/commit/7df1d16))
+
+
+### BREAKING CHANGES
+
+* **layout:** `session.selectedValue` is now `session.value`
+
+
+
+
 <a name="3.0.1"></a>
 ## [3.0.1](https://github.com/pie-framework/pie-ui/compare/@pie-ui/inline-choice@3.0.0...@pie-ui/inline-choice@3.0.1) (2018-04-24)